refactor: 이름 변경 모달 컴포넌트 추상화
