+++ frontend/src/layouts/SidebarLayout/components/SidebarEditUsernameModalButtonGroup/SidebarEditUsernameModalButtonGroup.styles.tsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+const StyledContainer = styled.div`
+  display: flex;
+  width: 40%;
+  gap: 2rem;
+`;
+
+export { StyledContainer };
